feat: add new vis components config file
