+++ web/src/components/chat/chat-content-components/VisComponents/RefsCard/index.tsx
